WIP last step of custom scenario
